msm: mdss: Don't allocate memory dynamically for small layer buffers
There's no reason to dynamically allocate memory for a single, small
struct instance (output_layer) when it can be allocated on the stack.
Additionally, layer_list's maximum size is defined, so that allocation
can be replaced with a static memory allocation instead. Note that the
static memory allocation is safe because this code does not run in
parallel.

Additionally, validate_info_list is limited by the maximum number of
layers allowed, so it can be replaced by a stack allocation.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Zile995 <stefan.zivkovic995@gmail.com>